requirements.txt
pyspark==3.5.1
pandas==2.2.2
pyyaml==6.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
pytest==8.3.3
pyarrow==16.1.0